+++ trunk/MetasyntacticShared/Classes/UIColor+Utilities.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#define RGBUIColor(r,g,b) [UIColor colorWithRed:(r)/255. green:(g)/255. blue:(b)/255. alpha:1]
-#define RGBAUIColor(r,g,b,a) [UIColor colorWithRed:(r)/255. green:(g)/255. blue:(b)/255. alpha:(a)/255.]
+#define RGBUIColor(r,g,b)    [UIColor colorWithRed:(CGFloat)((r)/255.) green:(CGFloat)((g)/255.) blue:(CGFloat)((b)/255.) alpha:1]
+#define RGBAUIColor(r,g,b,a) [UIColor colorWithRed:(CGFloat)((r)/255.) green:(CGFloat)((g)/255.) blue:(CGFloat)((b)/255.) alpha:(CGFloat)((a)/255.)]
 
 @interface UIColor(UIColorUtilities)
 - (CGFloat) red;
